[JAN-14] initialize project setup
