+++ .gitignore
@@ -1,21 +1,20 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-node_modules/
-.pnp/
+/node_modules
+/.pnp
 .pnp.js
-
-# testing
-coverage/
-
-# next.js
-.next/
-out/
-next-env.d.ts
 .idea/
 
+# testing
+/coverage
+
+# next.js
+/.next/
+/out/
+.next/
 # production
-build/
+/build
 
 # misc
 .DS_Store
@@ -27,16 +26,13 @@ yarn-debug.log*
 yarn-error.log*
 
 # local env files
-.env
 .env*.local
 
 # vercel
 .vercel
 
-# Turborepo
-.turbo
-
 # typescript
 *.tsbuildinfo
+next-env.d.ts
 
-.vscode/
+.env
