--- .prettierignore
@@ -1,4 +1,0 @@
-# Ignore artifacts:
-dist
-build
-node_modules/
