--- .prettierrc
@@ -1,9 +1,0 @@
-{
-    "bracketSpacing": true,
-    "printWidth": 80,
-    "singleQuote": true,
-    "trailingComma": "none",
-    "tabWidth": 2,
-    "useTabs": false,
-    "bracketSameLine": false
-}
