--- packages/clothes_gpt/.env.example
@@ -1,7 +1,0 @@
-POSTGRES_URL=
-POSTGRES_URL_NON_POOLING=
-POSTGRES_USER=
-POSTGRES_HOST=
-POSTGRES_PASSWORD=
-POSTGRES_DATABASE=
-POSTGRES_PRISMA_URL=
