--- packages/clothes_gpt/components/refresh-button.tsx
@@ -1,25 +1,0 @@
-'use client'
-
-import { useRouter } from 'next/navigation'
-import { useTransition } from 'react'
-
-export default function RefreshButton() {
-  const router = useRouter()
-  const [isPending, startTransition] = useTransition()
-
-  return (
-    <button
-      className={`${
-        isPending ? 'cursor-not-allowed text-gray-400' : ''
-      } text-sm text-gray-500 hover:text-gray-900`}
-      disabled={isPending}
-      onClick={() => {
-        startTransition(() => {
-          router.refresh()
-        })
-      }}
-    >
-      {isPending ? 'Refreshing...' : 'Refresh'}
-    </button>
-  )
-}
