--- packages/clothes_gpt/lib/utils.ts
@@ -1,8 +1,0 @@
-import ms from 'ms'
-
-export const timeAgo = (timestamp: Date, timeOnly?: boolean): string => {
-  if (!timestamp) return 'never'
-  return `${ms(Date.now() - new Date(timestamp).getTime())}${
-    timeOnly ? '' : ' ago'
-  }`
-}
