--- packages/clothes_gpt/next.config.js
@@ -1,11 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  experimental: {
-    appDir: true,
-  },
-  images: {
-    domains: ['pbs.twimg.com'],
-  },
-}
-
-module.exports = nextConfig
