--- packages/clothes_gpt/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "postgres-prisma",
-  "repository": "https://github.com/vercel/examples.git",
-  "license": "MIT",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "prisma generate && next dev",
-    "build": "prisma generate && prisma db push && prisma db seed && next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "prisma": {
-    "seed": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seed.ts"
-  },
-  "dependencies": {
-    "@prisma/client": "4.13.0",
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.37",
-    "@types/react-dom": "18.0.11",
-    "autoprefixer": "10.4.14",
-    "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
-    "ms": "^2.1.3",
-    "next": "13.3.2",
-    "postcss": "8.4.22",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "3.3.1",
-    "ts-node": "^10.9.1",
-    "typescript": "5.0.4"
-  },
-  "devDependencies": {
-    "@types/ms": "^0.7.31",
-    "prisma": "^4.13.0",
-    "turbo": "^1.9.3"
-  }
-}
