--- packages/clothes_gpt/tailwind.config.js
@@ -1,16 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    './pages/**/*.{js,ts,jsx,tsx}',
-    './components/**/*.{js,ts,jsx,tsx}',
-    './app/**/*.{js,ts,jsx,tsx}',
-  ],
-  theme: {
-    extend: {
-      fontFamily: {
-        default: ['var(--font-inter)'],
-      },
-    },
-  },
-  plugins: [],
-}
