--- packages/clothes_gpt/vercel.json
@@ -1,4 +1,0 @@
-{
-  "buildCommand": "pnpm turbo build",
-  "ignoreCommand": "pnpm dlx turbo-ignore"
-}
