+++ tsconfig.json
@@ -20,15 +20,12 @@
       }
     ],
     "paths": {
-      "@/*": ["./*"]
+      "@*": ["./*"],
     }
   },
   "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx",
-    ".next/types/**/*.ts",
-    "prisma/seed.ts"
-  ],
-  "exclude": ["node_modules"]
+    "next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "exclude": [
+    "./node_modules"
+  ]
 }
